small fixes

+++ core/src/me/shreyasr/arrow/util/PolarVelocity.java
@@ -2,10 +2,10 @@ package me.shreyasr.arrow.util;
 
 public class PolarVelocity {
 
-    public final double direction;
+    public final float direction;
     public final float speed;
 
-    public PolarVelocity(double direction, float speed) {
+    public PolarVelocity(float direction, float speed) {
         this.speed = speed;
         this.direction = direction;
     }
